add explicit cast for num_features

+++ jubatus/core/fv_converter/num_feature_impl.hpp
@@ -34,7 +34,7 @@ class num_value_feature : public num_feature {
   void add_feature(const std::string& key,
                    double value,
                    common::sfv_t& ret_fv) const {
-    ret_fv.push_back(std::make_pair(key, value));
+    ret_fv.push_back(std::make_pair(key, static_cast<float>(value)));
   }
 };
 
@@ -45,7 +45,7 @@ class num_string_feature : public num_feature {
                    common::sfv_t& ret_fv) const {
     std::stringstream ss;
     ss << key << "$" << value;
-    ret_fv.push_back(std::make_pair(ss.str(), 1.0));
+    ret_fv.push_back(std::make_pair(ss.str(), static_cast<float>(1.0)));
   }
 };
 
@@ -54,7 +54,9 @@ class num_log_feature : public num_feature {
   void add_feature(const std::string& key,
                    double value,
                    common::sfv_t& ret_fv) const {
-    ret_fv.push_back(std::make_pair(key, std::log(std::max(1.0, value))));
+    ret_fv.push_back(std::make_pair(
+        key,
+        static_cast<float>(std::log(std::max(1.0, value)))));
   }
 };
